data-augmentation-tool/image-augmentation-tool/requirements.txt
fastapi
uvicorn
python-multipart
numpy
pillow
scikit-image
